third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-date-expected.txt

CONSOLE WARNING: line 20: The specified value "text" does not conform to the required format, "yyyy-MM-dd".
This test ensures the correctness of Spatial Navigation (SNav) algorithm over date input element (having shadow DOM).
* Navigation steps:
1) Loads this page, focus goes to "start" automatically.
2) Focus moves away from input box, left and right to neighbor nodes and back.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Left		Right
PASS gFocusedDocument.activeElement.getAttribute("id") is "left"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "right"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS successfullyParsed is true

TEST COMPLETE
